feat: add @zarrita/indexing
