+++ packages/core/src/errors.ts
@@ -1,10 +1,3 @@
-export class NotImplementedError extends Error {
-	constructor(msg: string) {
-		super(msg);
-		this.name = "NotImplementedError";
-	}
-}
-
 export class NodeNotFoundError extends Error {
 	constructor(msg: string) {
 		super(msg);
@@ -12,29 +5,9 @@ export class NodeNotFoundError extends Error {
 	}
 }
 
-export class IndexError extends Error {
-	constructor(msg: string) {
-		super(msg);
-		this.name = "IndexError";
-	}
-}
-
 export class KeyError extends Error {
 	constructor(msg: string) {
 		super(msg);
 		this.name = "KeyError";
-	}
-}
-
-export class AssertionError extends Error {
-	constructor(msg: string) {
-		super(msg);
-		this.name = "ZarrAssertionError";
-	}
-}
-
-export function assert(condition: boolean, msg = "Assertion failed") {
-	if (!condition) {
-		throw new AssertionError(msg);
 	}
 }
