+++ packages/core/src/index.ts
@@ -1,11 +1,7 @@
-export * as ops from "./ops.js";
 export { KeyError, NodeNotFoundError } from "./errors.js";
 export { Array, Group, Location, root } from "./hierarchy.js";
 export { open } from "./open.js";
 export { create } from "./create.js";
 export { registry } from "./codecs.js";
-export { slice } from "./util.js";
-export type * from "./types.js";
+export { get_ctr, get_strides } from "./util.js";
 export type * from "./metadata.js";
-export { get as get_with_setter } from "./get.js";
-export { set as set_with_setter } from "./set.js";
