+++ packages/indexing/package.json
@@ -1,0 +1,21 @@
+{
+	"name": "@zarrita/indexing",
+	"type": "module",
+	"version": "0.0.0",
+	"sideEffects": false,
+	"scripts": {
+		"build": "tsc",
+		"clean": "rm -rf dist"
+	},
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.js"
+		}
+	},
+	"dependencies": {
+		"@zarrita/core": "workspace:^",
+		"@zarrita/storage": "workspace:^",
+		"@zarrita/typedarray": "workspace:^"
+	}
+}
