+++ packages/indexing/src/ops.ts
@@ -1,47 +1,31 @@
-// @ts-nocheck
 import type { Async, Readable, Writeable } from "@zarrita/storage";
-import type { Array } from "./hierarchy.js";
-import type {
-	Chunk,
-	GetOptions,
-	Indices,
-	SetOptions,
-	Slice,
-	TypedArray,
-} from "./types.js";
-import type {
-	BigintDataType,
-	DataType,
-	NumberDataType,
-	Scalar,
-} from "./metadata.js";
-
+import type * as core from "@zarrita/core";
 import { BoolArray } from "@zarrita/typedarray";
 import { get as get_with_setter } from "./get.js";
 import { set as set_with_setter } from "./set.js";
+import type { Indices, Slice, GetOptions, SetOptions, Projection } from "./types.js";
 
 // setting fns rely on some TypedArray apis not supported with our custom arrays
-
-type SupportedDataType = DataType;
+type SupportedDataType = core.NumberDataType | core.BigintDataType | core.Bool;
 
 export const setter = {
-	prepare<D extends DataType>(
-		data: TypedArray<D>,
+	prepare<D extends SupportedDataType>(
+		data: core.TypedArray<D>,
 		shape: number[],
 		stride: number[],
 	) {
 		return { data, shape, stride };
 	},
 	set_scalar<D extends SupportedDataType>(
-		dest: Chunk<D>,
+		dest: core.Chunk<D>,
 		sel: (number | Indices)[],
-		value: Scalar<D>,
+		value: core.Scalar<D>,
 	) {
 		set_scalar(compat(dest), sel, cast_scalar(dest, value));
 	},
 	set_from_chunk<D extends SupportedDataType>(
-		dest: Chunk<D>,
-		src: Chunk<D>,
+		dest: core.Chunk<D>,
+		src: core.Chunk<D>,
 		mapping: Projection[],
 	) {
 		set_from_chunk(compat(dest), compat(src), mapping);
@@ -54,42 +38,42 @@ export async function get<
 	Store extends Readable | Async<Readable>,
 	Sel extends (null | Slice | number)[],
 >(
-	arr: Array<D, Store>,
+	arr: core.Array<D, Store>,
 	selection: Sel | null = null,
 	opts: GetOptions<Parameters<Store["get"]>[1]> = {},
 ) {
-	return get_with_setter<D, Store, Chunk<D>, Sel>(arr, selection, opts, setter);
+	return get_with_setter<D, Store, core.Chunk<D>, Sel>(arr, selection, opts, setter);
 }
 
 /** @category Utility */
 export async function set<
 	D extends SupportedDataType,
 >(
-	arr: Array<D, (Readable & Writeable) | Async<Readable & Writeable>>,
+	arr: core.Array<D, (Readable & Writeable) | Async<Readable & Writeable>>,
 	selection: (null | Slice | number)[] | null,
-	value: Scalar<D> | Chunk<D>,
+	value: core.Scalar<D> | core.Chunk<D>,
 	opts: SetOptions = {},
 ) {
-	return set_with_setter<D, Chunk<D>>(arr, selection, value, opts, setter);
+	return set_with_setter<D, core.Chunk<D>>(arr, selection, value, opts, setter);
 }
 
 function compat<D extends SupportedDataType>(
-	arr: Chunk<D>,
-): Chunk<NumberDataType | BigintDataType> {
+	arr: core.Chunk<D>,
+): core.Chunk<D extends core.Bool ? core.Uint8 : D> {
 	// ensure strides are computed
 	return {
-		data: arr.data instanceof BoolArray
-			? (new Uint8Array(arr.data.buffer))
-			: arr.data,
+		// @ts-expect-error
+		data: arr.data instanceof BoolArray ? new Uint8Array(arr.data.buffer) : arr.data,
 		shape: arr.shape,
 		stride: arr.stride,
 	};
 }
 
-const cast_scalar = <D extends DataType>(
-	arr: Chunk<D>,
-	value: Scalar<D>,
-): Scalar<DataType> => {
+function cast_scalar<D extends core.DataType>(
+	arr: core.Chunk<D>,
+	value: core.Scalar<D>,
+): core.Scalar<D extends core.Bool ? core.Uint8 : D> {
+	// @ts-expect-error
 	if (arr.data instanceof BoolArray) return value ? 1 : 0;
 	return value as any;
 };
@@ -102,10 +86,10 @@ function indices_len(start: number, stop: number, step: number) {
 	return 0;
 }
 
-function set_scalar<D extends DataType>(
-	out: Pick<Chunk<D>, "data" | "stride">,
+function set_scalar<D extends core.NumberDataType | core.BigintDataType>(
+	out: Pick<core.Chunk<D>, "data" | "stride">,
 	out_selection: (Indices | number)[],
-	value: Scalar<D>,
+	value: core.Scalar<D>,
 ) {
 	if (out_selection.length === 0) {
 		out.data[0] = value;
@@ -115,7 +99,7 @@ function set_scalar<D extends DataType>(
 	const [curr_stride, ...stride] = out.stride;
 
 	if (typeof slice === "number") {
-		const data = out.data.subarray(curr_stride * slice) as TypedArray<D>;
+		const data = out.data.subarray(curr_stride * slice) as core.TypedArray<D>;
 		set_scalar({ data, stride }, slices, value);
 		return;
 	}
@@ -136,22 +120,14 @@ function set_scalar<D extends DataType>(
 	for (let i = 0; i < len; i++) {
 		const data = out.data.subarray(
 			curr_stride * (from + step * i),
-		) as TypedArray<D>;
+		) as core.TypedArray<D>;
 		set_scalar({ data, stride }, slices, value);
 	}
 }
 
-type Projection =
-	| { from: Indices; to: Indices }
-	| { from: null; to: number }
-	| {
-		from: number;
-		to: null;
-	};
-
-function set_from_chunk<D extends DataType>(
-	dest: Pick<Chunk<D>, "data" | "stride">,
-	src: Pick<Chunk<D>, "data" | "stride">,
+function set_from_chunk<D extends core.NumberDataType | core.BigintDataType>(
+	dest: Pick<core.Chunk<D>, "data" | "stride">,
+	src: Pick<core.Chunk<D>, "data" | "stride">,
 	projections: Projection[],
 ) {
 	const [proj, ...projs] = projections;
@@ -165,7 +141,7 @@ function set_from_chunk<D extends DataType>(
 		}
 		set_from_chunk(
 			{
-				data: dest.data.subarray(dstride * proj.to) as TypedArray<D>,
+				data: dest.data.subarray(dstride * proj.to) as core.TypedArray<D>,
 				stride: dstrides,
 			},
 			src,
@@ -180,7 +156,7 @@ function set_from_chunk<D extends DataType>(
 			return;
 		}
 		let view = {
-			data: src.data.subarray(sstride * proj.from) as TypedArray<D>,
+			data: src.data.subarray(sstride * proj.from) as core.TypedArray<D>,
 			stride: sstrides,
 		};
 		set_from_chunk(dest, view, projs);
