+++ packages/indexing/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "dist"
+	},
+	"include": ["index.ts", "src/**/*.ts", "__tests__/**/*.ts"]
+}
